+++ test/playwright/checkboxes.spec.ts
@@ -3,10 +3,14 @@ import {
   expectFormFailure,
   expectFormSuccess,
   expectSingleFieldValue,
+  checkAccessibility,
+  validateHtml,
 } from "./lib";
 
 test("checkbox", async ({ page }) => {
   await page.goto("/forms/checkbox/");
+  await checkAccessibility(page);
+  await validateHtml(page);
   await expect(await page.getByTestId("form")).toMatchAriaSnapshot(`
   - heading "Terms and conditions" [level=1]
   - checkbox "Terms and conditions I agree to terms and conditions"
@@ -15,6 +19,8 @@ test("checkbox", async ({ page }) => {
   await page.getByRole("button", { name: "Submit" }).click();
 
   await expectFormFailure(page);
+  await checkAccessibility(page);
+  await validateHtml(page);
   await expect(await page.getByTestId("form")).toMatchAriaSnapshot(`
   - alert:
     - heading "There is a problem" [level=2]
@@ -36,12 +42,17 @@ test("checkbox", async ({ page }) => {
   await page.getByRole("button", { name: "Submit" }).click();
 
   await expectFormSuccess(page);
+  await checkAccessibility(page);
+  await validateHtml(page);
   await expect(page.getByLabel("Terms and conditions")).toBeChecked();
   await expectSingleFieldValue(page, true);
 });
 
 test("checkbox-no-label", async ({ page }) => {
+  const ignoreAccessibiltyChecks = ["page-has-heading-one"];
   await page.goto("/forms/checkbox-no-label/");
+  await checkAccessibility(page, ignoreAccessibiltyChecks);
+  await validateHtml(page);
   await expect(await page.getByTestId("form")).toMatchAriaSnapshot(`
   - checkbox "I agree to terms and conditions"
   - text: I agree to terms and conditions
@@ -49,6 +60,8 @@ test("checkbox-no-label", async ({ page }) => {
   await page.getByRole("button", { name: "Submit" }).click();
 
   await expectFormFailure(page);
+  await checkAccessibility(page, ignoreAccessibiltyChecks);
+  await validateHtml(page);
   await expect(await page.getByTestId("form")).toMatchAriaSnapshot(`
   - alert:
     - heading "There is a problem" [level=2]
@@ -69,20 +82,24 @@ test("checkbox-no-label", async ({ page }) => {
   await page.getByRole("button", { name: "Submit" }).click();
 
   await expectFormSuccess(page);
+  await checkAccessibility(page, ignoreAccessibiltyChecks);
+  await validateHtml(page);
   await expect(page.getByLabel("Terms and conditions")).toBeChecked();
   await expectSingleFieldValue(page, true);
 });
 
 test("checkboxes", async ({ page }) => {
   await page.goto("/forms/checkboxes/");
+  await checkAccessibility(page);
+  await validateHtml(page);
   await expect(await page.getByTestId("form")).toMatchAriaSnapshot(`
   - group "Languages":
     - heading "Languages" [level=1]
     - paragraph: Select up to two programming languages
     - checkbox "C++"
     - text: C++
-    - checkbox "Python"
-    - text: Python
+    - checkbox "Python 3 ( Python 2 has been deprecated )"
+    - text: Python 3 ( Python 2 has been deprecated )
     - checkbox "PHP"
     - text: PHP
   - button "Submit"`);
@@ -90,6 +107,8 @@ test("checkboxes", async ({ page }) => {
   await page.getByRole("button", { name: "Submit" }).click();
 
   await expectFormFailure(page);
+  await checkAccessibility(page);
+  await validateHtml(page);
   await expect(page.getByRole("main")).toHaveText(/Select at least one item/);
   await expect(await page.getByTestId("form")).toMatchAriaSnapshot(`
   - alert:
@@ -103,8 +122,8 @@ test("checkboxes", async ({ page }) => {
     - paragraph: Select up to two programming languages
     - checkbox "C++"
     - text: C++
-    - checkbox "Python"
-    - text: Python
+    - checkbox "Python 3 ( Python 2 has been deprecated )"
+    - text: Python 3 ( Python 2 has been deprecated )
     - checkbox "PHP"
     - text: PHP
   - button "Submit"`);
@@ -117,6 +136,8 @@ test("checkboxes", async ({ page }) => {
   await page.getByRole("button", { name: "Submit" }).click();
 
   await expectFormFailure(page);
+  await checkAccessibility(page);
+  await validateHtml(page);
   await expect(page.getByRole("main")).toHaveText(
     /You must select no more than 2 items/,
   );
@@ -130,6 +151,8 @@ test("checkboxes", async ({ page }) => {
   await page.getByRole("button", { name: "Submit" }).click();
 
   await expectFormSuccess(page);
+  await checkAccessibility(page);
+  await validateHtml(page);
   await expect(page.getByLabel("C++")).toBeChecked();
   await expect(page.getByLabel("Python")).not.toBeChecked();
   await expect(page.getByLabel("PHP")).toBeChecked();